Bug 1747222 - [remote] Migrate TOP BROWSING CONTEXT to use custom UUID instead of browserId r=webdriver-reviewers,whimboo
We are not supporting the TOP_BROWSING_CONTEXT yet in BiDi, but message handler already supports it, and assumes browser ids. But as discussed we will need to use custom UUIDs, and on the side have a mapping between browser ids and UUIDs, as browserids might be available late.

Differential Revision: https://phabricator.services.mozilla.com/D135414

UltraBlame original commit: 23d2f0bfe24318574c464f0c062ce681e7a2df7b