Bug 1747176 - Make ClientWebGLContext::TexImage2D with a DOM element work with OffscreenCanvas. r=jgilbert
We cannot access DOM elements off the main thread, so any uses of that
must be a mistake. Also, if we are in use with an OffscreenCanvas, we
won't have an HTMLCanvasElement, so we need to make sure we pull the
nsIPrincipal from the OffscreenCanvas object instead.

Differential Revision: https://phabricator.services.mozilla.com/D136275

UltraBlame original commit: 6dcb3a75b392b9ef15a2c89fb34fe319b30433ac