Bug 1737969 - [devtools] Remove lodash findIndex/findLastIndex usages. r=jdescottes.
We take this opportunity to rename `removeInnerLocations` to `getInnerLocations`
and make it return an array of inner locations rather than returning the list
of original locations with inner locations stripped.
This makes the function easier to understand (not having to deal with indexes
+ splicing), and doesn't harm the only callsite where we can handle those inner
locations into the filter call that was done on the array returned from `removeInnerLocations`

The parser-worker was updated.

Differential Revision: https://phabricator.services.mozilla.com/D136095

UltraBlame original commit: 7c0b2b6e998d0d660b9e6ed111e570567a82a561