Bug 1750680: Don't use the role to check for iframes in AccessibleHandler. r=morgan
We need to handle OOP iframes specially when querying for IEnumVARIANT.
However, we can't rely on the role because it might be overridden by ARIA.
Instead, check for the absence of IAccessibleHyperText (which isn't supported for iframes) and a child count of 1.
This might trigger for a few things other than iframes, but it's not a problem to do what we're doing here whenever the child count is 1.

Differential Revision: https://phabricator.services.mozilla.com/D136257

UltraBlame original commit: 04afde5e03ca644419bc188fb07675695a16b3b0