Bug 1747371 - [devtools] Set the session context for the parent process target. r=nchevobbe
We were missing cached image request and probably CSP blocked request
from the parent process because of this.
We are spawning a NetworkEventContentWatcher for the parent process target.
That's semi-intentional. This might better be done via a parent process resource type,
like NETWORK_EVENT. But the current Resource framework doesn't allow to do that easily.

So that we currently spawn the NetworkEventContentWatcher (and NetworkEventStackTraceWatcher)
against the parent process target actor as a "FRAME" resource type.
This is handy as we need at least one of these two watchers for the parent process.

Differential Revision: https://phabricator.services.mozilla.com/D136325

UltraBlame original commit: 71af3a0623c43b1a17d7ca71b2d6c538883a7775