Bug 1740364 Part 2 - Tests for page-rule selectors r=emilio
Enable some existing tests that we can pass without pseudo-class parsing. This
leaves some selector tests disabled because they expect a paper-size with an
orientation to be serialized with the orientation even when the orientation is
'portrait', which we currently don't do.

Add some web-platform tests for page-rules with multiple selectors, and some
invalid selector tests.

One of the tests is currently disabled, as it appears that Servo allows a CSS
ident that begins with '--'.

Differential Revision: https://phabricator.services.mozilla.com/D135944

UltraBlame original commit: 358f7fa26c65a5f10da3326c40ad6b87030c1643