Bug 1734027 - Ensure we mark GLFeature::draw_buffers_indexed as unsupported if symbol load fails. r=jgilbert
When loading the symbols for draw_buffers_indexed, we marked it as for
draw_buffers. This would mean that if draw_buffers_indexed loading
failed, and we checked for the feature support, it would indicate it is
supported even though it is backed by nothing.

Differential Revision: https://phabricator.services.mozilla.com/D136381

UltraBlame original commit: 87b0596a08ab24bde54d6130f25bf64d85644954