Bug 1752063 - Remove -fno-common from js build. r=firefox-build-system-reviewers,andi
Incidentally, it has been the default since GCC 10 and clang 11, so this
is effectively a no-op when building with a modern compiler. For older
compilers, it doesn't matter that much anyways, because it was
cargo-culted from the Firefox build system, which removed it a long time
ago (bug 751511).

Differential Revision: https://phabricator.services.mozilla.com/D136980

UltraBlame original commit: a91d468530fbf1f341a25db7396582cc815ad08e