Bug 1715976 - Add Delazification helper thread task. r=arai
This patch adds delazification on helper threads. This patch is quite convoluted
as it includes all functions used to manipulate the DelazifyTask tasks.

A DelazifyTask is an HelperThread task which uses Parse threads to parse and
generate the bytecode of inner functions concurrently with the execution, and
which are scheduled after the initial parse is completed. DelazifyTask, as
opposed to other HelperThreadTask are tasks which are executed and re-scheduled
as long as inner function remain to be delazified based on the strategy used by
the DelazifyTask, implemented by a DelazifyStrategy.

A DelazifyStrategy is a class which holds ScriptIndex-es to be looked at, and
order them by priority or with a pre-defined order selected by the the
CompileOption.

As DelazifyTask-s are not attached to any object in the main thread, they are
automatically discarded when the source is no longer registered in the
StencilCache, or when the JSRuntime actively wait for the tasks to be collected
before the process shutdown.

Differential Revision: https://phabricator.services.mozilla.com/D136340

UltraBlame original commit: 67d0447ded80bece7e3b7f2492951fb63fbab76e