Bug 1571758 - Drop scroll method from SpatialNode. r=gw
It's not used for Gecko at all. And in the next commit we are going to have
multiple scroll offsets in a ScrollFrameInfo and encapsulate
ScrollFrameInfo.scroll_offset method so with this unused scroll function,
the next change will be quite messy.

Differential Revision: https://phabricator.services.mozilla.com/D133443

UltraBlame original commit: b13b738bc31b853dd2debba07b80dce8da4185ea