Bug 1750022 - Make tests in layout/base/tests/ dFPI compatible, r=smaug,anti-tracking-reviewers,timhuang
We are fixing mochitests that fail when network.cookie.cookieBehavior = 5, i.e. when we enable Total Cookie Protection.
This is most often due to the test assuming that an origin will always have access to its storage state when embedded as
a third party.

In this case I had to request storage access to allow an embedded iframe to use a BroadcastChannel to communicate with top-level pages of the same origin.
This was only a problem in the XOrigin case where the test html file is loaded in an embedded iframe and then can't communicate with the top-level window it created.

Differential Revision: https://phabricator.services.mozilla.com/D136608

UltraBlame original commit: 25b8169c550345845a36b04e76eeffb30752bb55