Bug 1750016 - Make tests in dom/tests/browser/ dFPI compatible, r=nika,timhuang
We are fixing mochitests that fail when network.cookie.cookieBehavior = 5, i.e. when we enable Total Cookie Protection.
This is most often due to the test assuming that an origin will always have access to its storage state when embedded as
a third party.

In this case I had to request first-party storage access for an embedded iframe. This test is using iframes not as an isolation mechanism,
but as way to open more windows and ensure storage persists between them. To test these codepaths first-party storage access
is required of the iframes.

Differential Revision: https://phabricator.services.mozilla.com/D136607

UltraBlame original commit: 5170692e884731a6bff522f5cd2d5a7903d1fef4