Bug 1750002 - Make tests in docshell/test/ dFPI compatible, r=smaug,anti-tracking-reviewers,timhuang
We are fixing mochitests that fail when network.cookie.cookieBehavior = 5, i.e. when we enable Total Cookie Protection.
This is most often due to the test assuming that an origin will always have access to its storage state when embedded as
a third party.

In this case I mostly had to request storage access to allow an embedded iframe to use a BroadcastChannel to communicate with top-level pages of the same origin.
For test `docshell/test/mochitest/test_bug1121701.html`, I also had to reorder the code slightly to make sure the test didn't end prematurely. This caused some
tabbing that makes the diff harder to read.

Differential Revision: https://phabricator.services.mozilla.com/D136602

UltraBlame original commit: 8277d1359adc8e9baa83715b42d550aa7483a17f