Bug 1753029 - [devtools] Add safe guard to avoid reparenting iframe nodeFront with wrong target. r=jdescottes.
This can be seen as a hot fix for Bug 1752342. Here we're not fixing the
root cause of the issue, but rather adding a protection mecanism so we don't
completely crash/freeze the inspector/the browser.
A follow-up bug will take care of Bug 1752342.

Differential Revision: https://phabricator.services.mozilla.com/D137450

UltraBlame original commit: 981f815e26f9d52a24296b81bc10c0cc47a6537b