Bug 1753131 - use fake device name changes to test devicechange events r=jib
"media.ondevicechange.fakeDeviceChangeEvent.enabled" does not change any
devices and so will not generate any events if/when no-op devicechange events
are filtered out.

The test is rearranged to reduce the wait for no event and run some steps in
parallel, so that the test runs in half the time.  This benefit will
accumulate as more scenarios are added to the test.

Differential Revision: https://phabricator.services.mozilla.com/D132906

UltraBlame original commit: eb741033c48d17868e68c1c1d718813324184e55